requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
httpx==0.25.2
pyjwt==2.8.0
python-dotenv==1.0.0
requests==2.31.0
python-dateutil>=2.8.0
pydantic==2.7.3
openpyxl==3.1.2
lark-oapi==1.3.6